remove superfluous autoreconnectonfail and fix SSL reconnect ...
i dont see a point in a separate  prefs.autoreconnectonfail
if we already have prefs.autoreconnect.
even worse, the setting was not reachable via gui and could
only be set via /SET.

as for the handler code for this error:
"error:1408F10B:SSL routines:SSL3_GET_RECORD:wrong version number"
"Are you sure this is a SSL capable server and port?"

the gtk frontend seems to ignore the return value of the function
that handles it (ssl_do_connect()), but in the text frontend
the timeout handler job was removed if the function returned 0.

now, if autoreconnect is set, it will reconnect on this error in
both frontends.