do not spam channel logs with repeated disconnects
when auto-connecting to a ZNC bouncer without entering the password
it usually makes a disconnect every 2 minutes.
this spams the log of every channel with disconnect messages.
to not repeat a disconnect message more than once, we check for
sess->total to see if the chan was opened on the previous connection.
if not, the disconnect message is suppressed.